0.4.0: added resource concept to core
api close output can be marked as a resource now, which can carry meta
data. this data can then be used by the origin.

this complements the origin concept, which gives the routing and
closing functions meta data about the origin of a request.

+ both the resource and the origin concept now require object type meta
data for a more homogenous usage in a possibly quite heterogenous
ecosystem.